+++ tests/toisto/test_app.py
@@ -7,6 +7,8 @@ from unittest.mock import MagicMock, Mock, patch
 import requests
 
 from toisto.metadata import VERSION
+from toisto.persistence.config import default_config
+from toisto.ui.text import CONFIG_LANGUAGE_TIP
 
 from ..base import ToistoTestCase
 
@@ -15,12 +17,13 @@ from ..base import ToistoTestCase
 class AppTest(ToistoTestCase):
     """Unit tests for the main method."""
 
-    concept_file_count = 0  # Counter for generating unique concept files
-
     def setUp(self):
         """Set up test fixtures."""
         super().setUp()
+        self.concept_file_count = 0  # Counter for generating unique concept files
+        self.current_version = Mock(json=Mock(return_value=[dict(name=VERSION)]))
         self.latest_version = Mock(json=Mock(return_value=[dict(name="v9999")]))
+        self.config = default_config()
 
     @patch("rich.console.Console.pager", MagicMock())
     @patch("toisto.persistence.spelling_alternatives.load_spelling_alternatives", Mock(return_value={}))
@@ -28,8 +31,10 @@ class AppTest(ToistoTestCase):
     @patch("builtins.input", Mock(side_effect=EOFError))
     @patch("pathlib.Path.exists", Mock(return_value=True))
     @patch("pathlib.Path.open")
-    def run_main(self, path_open: Mock) -> Mock:
+    @patch("toisto.app.read_config")
+    def run_main(self, read_config: Mock, path_open: Mock) -> Mock:
         """Run the main function and return the patched print method."""
+        read_config.return_value = self.config
         path_open.return_value.__enter__.return_value.read.side_effect = self.read_concept_file
         with patch("rich.console.Console.print") as patched_print, suppress(SystemExit):
             from toisto.app import main
@@ -37,8 +42,8 @@ class AppTest(ToistoTestCase):
             main()
         return patched_print
 
-    def read_concept_file(self):
-        """Generate a unique concept file."""
+    def read_concept_file(self) -> str:
+        """Return a unique concept file."""
         count = self.concept_file_count
         self.concept_file_count += 1
         return f"""{{
@@ -66,9 +71,9 @@ class AppTest(ToistoTestCase):
     @patch("requests.get")
     def test_current_version(self, requests_get: Mock) -> None:
         """Test that the practice command does not show the current version."""
-        requests_get.return_value = Mock(json=Mock(return_value=[dict(name=VERSION)]))
+        requests_get.return_value = self.current_version
         patched_print = self.run_main()
-        self.assertNotIn(VERSION, patched_print.call_args_list[3][0][0])
+        self.assertNotEqual(VERSION, patched_print.call_args_list[3][0][0].renderable)
 
     @patch.object(sys, "argv", ["toisto", "practice", "--target", "fi", "--source", "nl", "--file", "test"])
     @patch("requests.get")
@@ -85,3 +90,22 @@ class AppTest(ToistoTestCase):
         requests_get.return_value = self.latest_version
         patched_print = self.run_main()
         self.assertTrue(patched_print.call_args_list[2][0][0].title.startswith("Progress"))
+
+    @patch.object(sys, "argv", ["toisto", "practice", "--target", "fi", "--source", "nl"])
+    @patch("requests.get")
+    def test_language_configuration_tip(self, requests_get: Mock) -> None:
+        """Test that the practice command shows a tip when the user hasn't configured their languages."""
+        requests_get.return_value = self.current_version
+        patched_print = self.run_main()
+        self.assertEqual(CONFIG_LANGUAGE_TIP, patched_print.call_args_list[3][0][0].renderable)
+
+    @patch.object(sys, "argv", ["toisto", "practice"])
+    @patch("requests.get")
+    def test_no_language_configuration_tip(self, requests_get: Mock) -> None:
+        """Test that the practice command does not show a tip when the user has configured their languages."""
+        requests_get.return_value = self.current_version
+        self.config.add_section("languages")
+        self.config.set("languages", "target", "fi")
+        self.config.set("languages", "source", "nl")
+        patched_print = self.run_main()
+        self.assertNotIn(CONFIG_LANGUAGE_TIP, patched_print.call_args_list[3][0][0])